+++ analysis/src/SharedTypes.ml
@@ -311,12 +311,19 @@ end = struct
     {env with exported = structure.exported; pathRev; parent = Some env}
 end
 
+type typeArgContext = {
+  env: QueryEnv.t;
+  typeArgs: Types.type_expr list;
+  typeParams: Types.type_expr list;
+}
+
 type polyVariantConstructor = {
   name: string;
   displayName: string;
   args: Types.type_expr list;
 }
 
+(* TODO(env-stuff) All envs for bool string etc can be removed. *)
 type innerType = TypeExpr of Types.type_expr | ExtractedType of completionType
 and completionType =
   | Tuple of QueryEnv.t * Types.type_expr list * Types.type_expr
@@ -337,8 +344,6 @@ and completionType =
       constructors: Constructor.t list;
       variantDecl: Types.type_declaration;
       variantName: string;
-      typeArgs: Types.type_expr list;
-      typeParams: Types.type_expr list;
     }
   | Tpolyvariant of {
       env: QueryEnv.t;
@@ -780,10 +785,11 @@ module Completion = struct
     docstring: string list;
     kind: kind;
     detail: string option;
+    typeArgContext: typeArgContext option;
   }
 
-  let create ~kind ~env ?(docstring = []) ?filterText ?detail ?deprecated
-      ?insertText name =
+  let create ~kind ~env ?typeArgContext ?(docstring = []) ?filterText ?detail
+      ?deprecated ?insertText name =
     {
       name;
       env;
@@ -795,10 +801,11 @@ module Completion = struct
       insertTextFormat = None;
       filterText;
       detail;
+      typeArgContext;
     }
 
-  let createWithSnippet ~name ?insertText ~kind ~env ?sortText ?deprecated
-      ?filterText ?detail ?(docstring = []) () =
+  let createWithSnippet ~name ?typeArgContext ?insertText ~kind ~env ?sortText
+      ?deprecated ?filterText ?detail ?(docstring = []) () =
     {
       name;
       env;
@@ -810,6 +817,7 @@ module Completion = struct
       insertTextFormat = Some Protocol.Snippet;
       filterText;
       detail;
+      typeArgContext;
     }
 
   (* https://microsoft.github.io/language-server-protocol/specifications/specification-current/#textDocument_completion *)